+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: lwourms <lwourms@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/11 15:21:17 by lwourms           #+#    #+#             */
-/*   Updated: 2020/12/11 15:50:13 by lwourms          ###   ########lyon.fr   */
+/*   Updated: 2021/01/02 15:57:48 by lwourms          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_isprint(int c);
 int		ft_isalpha(int c);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
-int		ft_strlen(char *str);
+size_t	ft_strlen(char *str);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
